scripts/requirements-depth.txt
onnxruntime>=1.17.0
Pillow>=10.0.0
numpy>=1.26.0
pyyaml>=6.0
